+++ src/Contracts/ProvinceSeeder.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Konekt\Address\Contracts;
+
+interface ProvinceSeeder
+{
+    public static function getCountryCode(): string;
+
+    /**
+     * @return array|ProvinceType[]
+     */
+    public static function getProvinceTypes(): array;
+
+}