+++ src/resources/database/seeds/ProvincesAndRegionsOfBelgium.php
@@ -15,9 +15,16 @@ declare(strict_types=1);
 namespace Konekt\Address\Seeds;
 
 use Illuminate\Database\Seeder;
+use Konekt\Address\Contracts\ProvinceSeeder;
+use Konekt\Address\Models\ProvinceType;
 
-class ProvincesAndRegionsOfBelgium extends Seeder
+class ProvincesAndRegionsOfBelgium extends Seeder implements ProvinceSeeder
 {
+    use IsProvinceSeeder;
+
+    protected static string $forCountry = 'BE';
+    protected static array $provinceTypes = [ProvinceType::REGION, ProvinceType::PROVINCE];
+
     public function run()
     {
         \DB::table('provinces')->insert([